Rename prolog 'ops' package to 'dsl'
